Refactor shared targeting code
