+++ packages/server/src/api/ampEpicRouter.ts
@@ -4,11 +4,9 @@ import cors from 'cors';
 import { AmountsTests, AmountsCardData, OneOffSignupRequest } from '@sdc/shared/dist/types';
 import fetch from 'node-fetch';
 import {
-    buildAmpEpicCampaignCode,
     buildReminderFields,
     countryCodeToCountryGroupId,
     getLocalCurrencySymbol,
-    isSupportUrl,
 } from '@sdc/shared/dist/lib';
 import { getAmpVariantAssignments } from '../lib/ampVariantAssignments';
 import { ampEpic } from '../tests/amp/ampEpic';
@@ -17,6 +15,9 @@ import { ValueProvider } from '../utils/valueReloader';
 import { TickerDataProvider } from '../lib/fetchTickerData';
 import { AmpEpicTest } from '../tests/amp/ampEpicModels';
 import { OphanComponentEvent } from '@guardian/libs';
+import { buildAmpEpicCampaignCode } from '../lib/tracking';
+
+const isSupportUrl = (baseUrl: string): boolean => /\bsupport\./.test(baseUrl);
 
 export const setOneOffReminderEndpoint = (): string =>
     isProd
