--- packages/shared/src/factories/epicVariant.ts
@@ -1,21 +1,0 @@
-import { EpicVariant } from '@sdc/shared/types';
-import { Factory } from 'fishery';
-
-export default Factory.define<EpicVariant>(() => ({
-    name: 'Example Variant',
-    heading: 'Since you’re here...',
-    paragraphs: [
-        '... we have a small favour to ask. More people, like you, are reading and supporting the Guardian’s independent, investigative journalism than ever before. And unlike many news organisations, we made the choice to keep our reporting open for all, regardless of where they live or what they can afford to pay.',
-        'The Guardian will engage with the most critical issues of our time – from the escalating climate catastrophe to widespread inequality to the influence of big tech on our lives. At a time when factual information is a necessity, we believe that each of us, around the world, deserves access to accurate reporting with integrity at its heart.',
-        'Our editorial independence means we set our own agenda and voice our own opinions. Guardian journalism is free from commercial and political bias and not influenced by billionaire owners or shareholders. This means we can give a voice to those less heard, explore where others turn away, and rigorously challenge those in power.',
-        'We hope you will consider supporting us today. We need your support to keep delivering quality journalism that’s open and independent. Every reader contribution, however big or small, is so valuable. ',
-    ],
-    highlightedText:
-        'Support The Guardian from as little as %%CURRENCY_SYMBOL%%1 - and it only takes a minute. Thank you.',
-    backgroundImageUrl:
-        'https://images.unsplash.com/photo-1494256997604-768d1f608cac?ixlib=rb-1.2.1&auto=format&fit=crop&w=1701&q=80',
-    cta: {
-        text: 'Contribute to The Guardian!',
-        baseUrl: 'https://support.theguardian.com/support',
-    },
-}));
