--- packages/shared/src/factories/index.ts
@@ -1,14 +1,0 @@
-import { register } from 'fishery';
-import targeting from './targeting';
-import test from './test';
-import { pageTracking, testTracking, tracking } from './tracking';
-import epicVariant from './epicVariant';
-
-export const factories = register({
-    pageTracking,
-    targeting,
-    test,
-    testTracking,
-    tracking,
-    epicVariant,
-});
