--- packages/shared/src/factories/targeting.ts
@@ -1,14 +1,0 @@
-import { EpicTargeting } from '@sdc/shared/types';
-import { Factory } from 'fishery';
-
-export default Factory.define<EpicTargeting>(() => ({
-    contentType: 'Article',
-    sectionId: 'culture',
-    shouldHideReaderRevenue: false,
-    isMinuteArticle: false,
-    isPaidContent: false,
-    isRecurringContributor: false,
-    tags: [],
-    showSupportMessaging: true,
-    hasOptedOutOfArticleCount: false,
-}));
