+++ packages/shared/src/lib/index.ts
@@ -1,6 +1,5 @@
 export * from './geolocation';
 export * from './placeholders';
 export * from './reminderFields';
-export * from './tracking';
 export * from './history';
 export * from './viewLog';
